refactor: move storage manager to its own module
